consumer/requirements.txt
kafka-python
mysql-connector-python